performance-optimization-sql/requirements.txt
# requirements.txt
# Python dependencies for the SQL Performance Optimization Pipeline

psycopg2-binary  # For connecting and executing commands against PostgreSQL
pandas           # Useful for data handling and structured output
matplotlib       # Library for creating the performance comparison graph